refactored :multiselect_muxer_interface
